Fix segfault in attribute retrival
When attempting to retrieve udev attributes from a hash a crash was
happening when using 'hv_iterkey'. Switched this to use 'hv_iterkeysv'
to extract the hash key as a scalar to save having to store length
values. This resolves the segfault, but still unsure what it is about
using storing value lengths that has been causing the trouble.

Tidied up the surrounding code so that attribute retrieval works in
logicial stages now.