Added udev device property filtering
Device properties can now be provided in the search parameters when
retrieving a list of matching devices. This is provided via a hash
reference called 'property' with the search properties.

Included an example in the test script for this now too.

Also removed the hardcoded property search as filtering for properties
now works at the user level.

Change-Id: I035347b55e77e44fe7636f45b665213c2a16ba4b
BUG-ID:  N/A